+++ CHANGELOG.md
@@ -1,16 +1,26 @@
-## 1.0.0
+# 1.0.0
 
 * Renaming members for better intuitivity
 * Adding scoped `Scope`, so features can be scoped and disposed when the `Scope` is disposed
 
-## 1.0.1
+### 1.0.1
 
 * Improving logging: support for external logging and simpler log messages
 
-## 1.0.2
+### 1.0.2
 
 * Fixed logging on component changes
 
-## 1.0.3
+### 1.0.3
 
 * Fixed wrong class name (FlowFeatureDisposalLogic)
+
+### 1.0.4
+
+* Fixed bug in FlowScope getComponent
+* Added getComponentOrNull
+
+## 1.1.0
+
+* Added support for `dispose` in all types of Logic (it will be called once a `FlowFeature` is disposed)
+  This allows, for example, to get a `StreamSubscription<T>` and then close it when it is no longer needed.